# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 推しスケジュール API
  version: 1.0.0
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Error:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties: { code: { type: string }, message: { type: string } }
        requestId: { type: string }
    Handle:
      type: object
      required: [handle]
      properties: { handle: { type: string, pattern: '^@[A-Za-z0-9._-]{3,30}$' } }
security: [{ bearerAuth: [] }]
paths:
  /health:
    get:
      security: []
      responses: { '200': { description: healthy } }
  /api/v1/me:
    get:
      responses: { '200': { description: current user }, '401': { description: unauthorized } }
  /api/v1/onboarding:
    post:
      responses: { '200': { description: configured }, '403': { description: not invited } }
  /api/v1/channels:
    get: { responses: { '200': { description: subscriptions } } }
    post:
      {
        responses:
          {
            '201': { description: created },
            '409': { description: duplicate },
            '422': { description: limit },
          },
      }
  /api/v1/channels/resolve:
    post:
      requestBody:
        {
          required: true,
          content: { application/json: { schema: { $ref: '#/components/schemas/Handle' } } },
        }
      responses: { '200': { description: resolved channel }, '404': { description: not found } }
  /api/v1/channels/{subscriptionId}:
    parameters: [{ in: path, name: subscriptionId, required: true, schema: { type: string } }]
    patch: { responses: { '200': { description: updated }, '404': { description: not found } } }
    delete: { responses: { '204': { description: deleted }, '404': { description: not found } } }
  /api/v1/channels/{subscriptionId}/sync:
    parameters: [{ in: path, name: subscriptionId, required: true, schema: { type: string } }]
    post:
      {
        responses:
          {
            '202': { description: accepted },
            '409': { description: running },
            '429': { description: cooldown },
          },
      }
  /api/v1/sync-status:
    get: { responses: { '200': { description: latest status } } }
  /api/v1/google/reconnect:
    post: { responses: { '200': { description: reconnected } } }
  /api/v1/account:
    delete:
      { responses: { '202': { description: deletion started }, '204': { description: deleted } } }